Error moved to common and fixed error page layout
