+++ common-ui/components/error/error.styles.tsx
