+++ terminology-ui/src/pages/404.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Error from '@app/common/components/error/error';
+import Error from 'yti-common-ui/error/error';
 import ErrorLayout from 'yti-common-ui/layout/error-layout';
 import {
   CommonContextProvider,
